refactor: split accounts controller into files and add input validation
