+++ src/controllers/accounts/index.ts
@@ -1,0 +1,5 @@
+export * from './get-accounts';
+export * from './get-account-by-id';
+export * from './create-account';
+export * from './update-account';
+export * from './delete-account';
