+++ src/routes/accounts.route.ts
@@ -3,17 +3,22 @@ import { authenticateJwt } from '@middlewares/passport';
 import {
   getAccounts,
   getAccountById,
+  getAccountByIdSchema,
   createAccount,
+  createAccountSchema,
   updateAccount,
+  updateAccountSchema,
   deleteAccount,
-} from '@controllers/accounts.controller';
+  deleteAccountSchema,
+} from '@controllers/accounts';
+import { validateEndpoint } from '@middlewares/validations';
 
 const router = Router({});
 
 router.get('/', authenticateJwt, getAccounts);
-router.get('/:id', authenticateJwt, getAccountById);
-router.post('/', authenticateJwt, createAccount);
-router.put('/:id', authenticateJwt, updateAccount);
-router.delete('/:id', authenticateJwt, deleteAccount);
+router.get('/:id', authenticateJwt, validateEndpoint(getAccountByIdSchema), getAccountById);
+router.post('/', authenticateJwt, validateEndpoint(createAccountSchema), createAccount);
+router.put('/:id', authenticateJwt, validateEndpoint(updateAccountSchema), updateAccount);
+router.delete('/:id', authenticateJwt, validateEndpoint(deleteAccountSchema), deleteAccount);
 
 export default router;
